Moved show_in_explorer to package, change var names, re #24812
