+++ qt/python/mantidqt/widgets/codeeditor/tab_widget/codeeditor_tab_presenter.py
@@ -8,26 +8,9 @@
 from __future__ import absolute_import
 
 import os
-import subprocess
-import sys
 
 from mantid.kernel import logger
-
-
-class ShowInExplorer(object):
-    @staticmethod
-    def open_directory(path):
-        if sys.platform == "win32":
-            os.startfile(path)
-        elif sys.platform == 'darwin':
-            subprocess.check_call(['open', '--', path])
-        elif sys.platform == 'linux2':
-            call_params = ['xdg-open', path]
-            try:
-                subprocess.check_call(call_params)
-            except subprocess.CalledProcessError as ex:
-                logger.notice("Could not manage to open the folder in explorer.")
-                logger.debug("Error encountered: {}".format(ex))
+from mantidqt.utils.show_in_explorer import ShowInExplorer
 
 
 class CodeEditorTabPresenter(object):
