+++ qt/python/mantidqt/widgets/codeeditor/tab_widget/test/test_codeeditor_tab_presenter.py
@@ -1,0 +1,15 @@
+# Mantid Repository : https://github.com/mantidproject/mantid
+#
+# Copyright &copy; 2019 ISIS Rutherford Appleton Laboratory UKRI,
+#     NScD Oak Ridge National Laboratory, European Spallation Source
+#     & Institut Laue - Langevin
+# SPDX - License - Identifier: GPL - 3.0 +
+#  This file is part of the mantidqt package
+from __future__ import absolute_import
+
+from unittest import TestCase
+
+
+class CodeEditorTabPresenterTest(TestCase):
+    def test_action_show_in_explorer(self):
+        self.fail()
